Generalize Model on trait implementing Material
Previously a Material enum was used to handle generalizing over types
that implement MaterialResponse. This is unnecessary if Model takes a
type bound on Material. Ultimately Model should only ever be referenced
as a type or trait object with Renderable, which helps with Model<G,M>
beginning to get unweildy.

- Remove Material enum
- Rename MaterialResponse to Material
- ModelList is now Vec<Box<Renderable>> so it can hold objects with
different materials